Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,6 +3,6 @@ int main()
 {
     double num;
     scanf("%lf",&num);
-    printf("You entered: %.2lf",num);
+    printf("You entered: %.2lf",(double)num);
     return 0;
 }
